# apps/api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/catalog/products/{slug}": {
      "get": {
        "operationId": "CatalogController_product",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailDto"
                }
              }
            }
          }
        },
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/admin/products": {
      "post": {
        "operationId": "AdminProductController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-catalog"
        ]
      },
      "get": {
        "operationId": "AdminProductController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-catalog"
        ]
      }
    },
    "/admin/products/{id}": {
      "get": {
        "operationId": "AdminProductController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-catalog"
        ]
      },
      "patch": {
        "operationId": "AdminProductController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-catalog"
        ]
      },
      "delete": {
        "operationId": "AdminProductController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-catalog"
        ]
      }
    },
    "/admin/products/{id}/toggle-active": {
      "patch": {
        "operationId": "AdminProductController_toggleActive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-catalog"
        ]
      }
    },
    "/admin/products/bulk": {
      "post": {
        "operationId": "AdminProductController_bulk",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkProductActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-catalog"
        ]
      }
    },
    "/admin/products/import": {
      "post": {
        "operationId": "AdminProductController_importCsv",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ImportProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-catalog"
        ]
      }
    },
    "/admin/products/{id}/images": {
      "post": {
        "operationId": "AdminProductController_addImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddProductImageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductImageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-catalog"
        ]
      }
    },
    "/admin/products/{id}/images/{imageId}": {
      "delete": {
        "operationId": "AdminProductController_removeImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductImageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-catalog"
        ]
      }
    },
    "/admin/products/{id}/images/reorder": {
      "patch": {
        "operationId": "AdminProductController_reorderImages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderProductImagesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-catalog"
        ]
      }
    },
    "/admin/products/{id}/variants": {
      "post": {
        "operationId": "AdminProductVariantController_create",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductVariantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductVariantResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-catalog"
        ]
      }
    },
    "/admin/products/{id}/variants/{variantId}": {
      "patch": {
        "operationId": "AdminProductVariantController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductVariantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductVariantResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-catalog"
        ]
      },
      "delete": {
        "operationId": "AdminProductVariantController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductVariantResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-catalog"
        ]
      }
    },
...
  },
  "info": {
    "title": "E-commerce API",
    "description": "API para tienda y panel administrativo",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CategorySummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "metaTitle": {
            "type": "string"
          },
          "metaDescription": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "ProductVariantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sku": {
            "type": "string"
          },
          "size": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "garmentType": {
            "type": "string"
          },
          "stockMode": {
            "type": "string"
          },
          "productionLeadTimeDays": {
            "type": "number"
          },
          "priceAdjustment": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          },
          "reservedQuantity": {
            "type": "number"
          },
          "availableQuantity": {
            "type": "number"
          },
          "inStock": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "sku",
          "stockMode"
        ]
      },
...
      "ProductDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "metaTitle": {
            "type": "string"
          },
          "metaDescription": {
            "type": "string"
          },
          "basePrice": {
            "type": "number"
          },
          "displayPrice": {
            "type": "number"
          },
          "category": {
            "$ref": "#/components/schemas/CategorySummaryDto"
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariantDto"
            }
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductImageDto"
            }
          },
          "averageRating": {
            "type": "number"
          },
          "reviewCount": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "basePrice",
          "displayPrice",
          "variants",
          "images",
          "createdAt"
        ]
      },
...
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "description": {
            "type": "string"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "imageAssetId": {
            "type": "string",
            "format": "uuid"
          },
          "basePrice": {
            "type": "number",
            "minimum": 0
          },
          "isActive": {
            "type": "boolean"
          },
          "metaTitle": {
            "type": "string",
            "maxLength": 200
          },
          "metaDescription": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "name",
          "slug",
          "basePrice"
        ]
      },
...
      "ProductVariantResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sku": {
            "type": "string"
          },
          "size": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "garmentType": {
            "type": "string"
          },
          "stockMode": {
            "type": "string"
          },
          "productionLeadTimeDays": {
            "type": "number"
          },
          "priceAdjustment": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sku",
          "stockMode",
          "productionLeadTimeDays",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProductResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "metaTitle": {
            "type": "string"
          },
          "metaDescription": {
            "type": "string"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "basePrice": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/ProductCategorySummaryDto"
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariantResponseDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "basePrice",
          "isActive",
          "createdAt",
          "updatedAt",
          "variants"
        ]
      },
      "AdminProductListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductResponseDto"
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "number"
              },
              "limit": {
                "type": "number"
              },
              "total": {
                "type": "number"
              }
            },
            "required": [
              "page",
              "limit",
              "total"
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "AdminProductDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "metaTitle": {
            "type": "string"
          },
          "metaDescription": {
            "type": "string"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "basePrice": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/ProductCategorySummaryDto"
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariantResponseDto"
            }
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductImageDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "basePrice",
          "isActive",
          "createdAt",
          "updatedAt",
          "variants",
          "images"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "description": {
            "type": "string"
          },
          "categoryId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "basePrice": {
            "type": "number",
            "minimum": 0
          },
          "isActive": {
            "type": "boolean"
          },
          "metaTitle": {
            "type": "string",
            "maxLength": 200
          },
          "metaDescription": {
            "type": "string",
            "maxLength": 500
          }
        }
      },
      "BulkProductActionDto": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "delete",
              "activate",
              "deactivate"
            ]
          },
          "ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "action",
          "ids"
        ]
      },
      "ImportProductDto": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "format": "binary"
          }
        },
        "required": [
          "file"
        ]
      },
      "AddProductImageDto": {
        "type": "object",
        "properties": {
          "assetId": {
            "type": "string",
            "format": "uuid"
          },
          "productVariantId": {
            "type": "string",
            "format": "uuid"
          },
          "altText": {
            "type": "string"
          },
          "sortOrder": {
            "type": "number"
          }
        },
        "required": [
          "assetId"
        ]
      },
      "ProductImageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "assetId": {
            "type": "string"
          },
          "productVariantId": {
            "type": "string"
          },
          "altText": {
            "type": "string"
          },
          "sortOrder": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "assetId",
          "sortOrder"
        ]
      },
      "ReorderProductImagesDto": {
        "type": "object",
        "properties": {
          "imageIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "imageIds"
        ]
      },
      "CreateProductVariantDto": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string"
          },
          "size": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "garmentType": {
            "type": "string"
          },
          "stockMode": {
            "type": "string",
            "enum": [
              "MADE_TO_ORDER",
              "TRACKED"
            ]
          },
          "productionLeadTimeDays": {
            "type": "number",
            "minimum": 0
          },
          "priceAdjustment": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "sku"
        ]
      },
      "UpdateProductVariantDto": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string"
          },
          "size": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "garmentType": {
            "type": "string"
          },
          "stockMode": {
            "type": "string",
            "enum": [
              "MADE_TO_ORDER",
              "TRACKED"
            ]
          },
          "productionLeadTimeDays": {
            "type": "number",
            "minimum": 0
          },
          "priceAdjustment": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}